PR Response: Move methods into utils.py.
